Custom the save method of the Rule model to handle the annotation "rule"
When creating or updating a rule on Promgen, the corresponding rule data on Prometheus will be given an annotation rule that is the link to that rule on Promgen. However, the annotation rule on Promgen always has the URI "/rule/0" because it has not been intentionally set up.

To ensure data is synced between the two applications, we customize the save method of the Rule model.