Add "project" field to the Farm model
We have planned to make Farm a child object of Project. Each Project will have its
own Farm and will not share it with other Projects. This means the real relationship
between a Project and a Farm will be one-to-one. Therefore, we should use Django's
one-to-one relationship to implement the relationship between the Farm and Project
models.

In the official Django documentation, one-to-one relationships are asymmetrical, and the
definition of the parent should be placed in the child. However, we have currently
defined "farm" as a field of the Project model. Therefore, we will switch the direction
by adding a "project" field to the Farm model, populate data into the
promgen_farm.project_id column, and then remove the "farm" field in the Project model.
Before removing the "farm" field, both definitions will have related_name="+" so that
the reverse relationships do not clash with the other field.

Ref: https://docs.djangoproject.com/en/4.2/topics/db/examples/one_to_one/