Change the way accesses a Project from a Farm
Due to changing relationship between Farm and Project to one-to-one, the way we access
the parent Project from a Farm will be "farm.project" instead of "farm.project_set".
We have modified all necessary code changes correspond to this.
A signal also became redundant and has been removed because we have defined
on_delete=models.CASCADE in the Farm model class.